requirements.txt
fastapi
uvicorn
langchain
langchain-openai
openai
pydantic
pydantic-settings
python-dotenv